Web Scrapping/ChatGPT/requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
httpx>=0.24.0

